[Digsite] Winches
Fixed ropes not getting taken on use on winch.

Fixed a bug where multiple scenarios worth of text and player dialog print happen when climbing down the winch. (Ex. Rope it attached so when climbing down, successful climb text and no rope attached text print)

When a player can't use the winch, if a workman has to be spawned because there wasn't one close enough to teleport to the player, the workman is removed after he speaks to the player.

General code cleanup.